fix: Check variant platform before deletion/secretReset/findById
